Renamed RelativeDateCheck to AfterRelativeDate
