+++ Packs/CommonScripts/ReleaseNotes/1_2_62.md
@@ -1,4 +1,4 @@
 
 #### Scripts
-##### New: BeforeRelativeDate
-- Evaluates to True if the left value is sooner than the right date.
+##### New: AfterRelativeDate
+- Added a new filter which checks that given time has happend after relative time
